// src/app/api/models/alumno-registro-request.ts
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

export interface AlumnoRegistroRequest {
  direccion?: string;
  dni?: string;
  emailTutor?: string;
  grado?: string;
  nivel?: string;
  nombreCompleto?: string;
  responsablePago?: string;
  seccion?: string;
  telefonoTutor?: string;
}