// src/app/api/models/alumno-response.ts
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

export interface AlumnoResponse {
  codAlumno?: string;
  direccion?: string;
  dni?: string;
  emailTutor?: string;
  grado?: string;
  id?: number;
  nivel?: string;
  nombreCompleto?: string;
  responsablePago?: string;
  seccion?: string;
  telefonoTutor?: string;
}